Memoize the PredictionIO clients
The official predictionio ruby sdk is written in such a way that you
must always use the same instance of a client. Failure to do so results
in a thread per client which will enter a sleep_forever state since
these threads do not idle out and will wait for new packages. In
instances where a worker thread is processing jobs such as in Sidekiq
you can find yourself with many jobs instantiating clients. By memoizing
the client we are reducing the rate of instantiation of clients however,
you may need to store your engine instance in a location which is
accessible or pass a reference to it constantly.